Major refactor of packaging functionality; allows for - filesystem-syncing w/ meaningful filenames, - much cleaner json package files/processing - history tracking in UpdateLog records
